+++ Contentstack.Core/Models/Query.cs
@@ -1,4 +1,4 @@
-﻿using Newtonsoft.Json.Linq;
+using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -1021,7 +1021,7 @@ namespace Contentstack.Core.Models
 
         /// <summary>
         /// Request specific asset-related metadata in the response (CDA asset_fields[]).
-        /// Supported values include: user_defined_fields, embedded, ai_suggested, visual_markups.
+        /// Valid parameters: user_defined_fields, embedded_metadata, ai_generated_metadata, visual_markups.
         /// </summary>
         /// <param name="fields">Asset field names to include.</param>
         /// <returns>Current instance of Query for chaining.</returns>